+++ app/controllers/api/v1/projects_controller.rb
@@ -21,4 +21,8 @@ class Api::V1::ProjectsController < Api::V1::BaseController
   # # GET /api/v1/projects/1-AUTHENTICATION_TOKEN/randomizations.json
   # def randomizations
   # end
+
+  # # GET /api/v1/projects/1-AUTHENTICATION_TOKEN/expression.json
+  # def expression
+  # end
 end