+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### Enhancements
 - **API Changes**
   - Randomization reports now display months in descending format
+  - Add support for reports that filter based on an expression, group by a
+    specified date, and then list results by month and by site
 - **Export Changes**
   - New adverse event module sheets are now exportable
 - **Team Changes**